fix/rename components
