+++ src/app/App.tsx
@@ -1,4 +1,4 @@
-import MainPage from '../pages/MainPage/MainPage';
+import MainPage from '../pages/MainPage';
 
 function App() {
   return (
