+++ src/widgets/Accordeon/styles.module.scss
